"Надувание" макета активити NotesListFragment(ом)
