+++ app/src/main/res/layout/activity_notes_list.xml
@@ -8,8 +8,8 @@
     android:orientation="vertical"
     tools:context=".ui.NotesListActivity">
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/list"
+    <FrameLayout
+        android:id="@+id/list_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
